[AIRFLOW-XXXX] Move Big Table sensor to sensors package
